100DaysofCode/Intermediate/Day-49-Automating Job Applications on Linkedin/requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
